# PERF: Improve performance when accessing GroupBy.groups